Add Discount calculator Logic

+++ ShoppingCalculator/Bill.swift
@@ -10,15 +10,15 @@ import Foundation
 struct Bill: Equatable {
     let items: [Product]
 
-    var subTotal: Decimal {
+    var subTotal: Float {
         items.reduce(0) { $0 + $1.total }
     }
 
-    var grocesrySubTotal: Decimal {
+    var grocesrySubTotal: Float {
         items.filter { $0.type.isGrocery }.reduce(0) { $0 + $1.total }
     }
 
-    var nonGrocerySubTotal: Decimal {
+    var nonGrocerySubTotal: Float {
         items.filter { !$0.type.isGrocery }.reduce(0) { $0 + $1.total }
     }
 }
